+++ CHANGELOG.md
@@ -7,8 +7,12 @@
 # CHANGELOG
 
 <p>
+<b>Version 2.1.37</b> - July 2023<br/>
+- Load control: added msg.shedding to force shed/unshed.<br/>
+</p>
+<p>
 <b>Version 2.1.36</b> - July 2023<br/>
-- Bugfix: in some cases, the eventsource lost the connection while in idle for more than 15 minutes; fixed.<br/>
+- HUE Server bugfix: in some cases, the eventsource lost the connection while in idle for more than 15 minutes; fixed.<br/>
 </p>
 <p>
 <b>Version 2.1.35</b> - July 2023<br/>